feat:  Add new Remap interface to relations (#71)
Remap() is an alternative to the current Remap variants currently found
on the Builder interface. It preserves the existing behavior including
checking the proposed mapping for validity. It also modifies the
relation instead of returning a copy. All of the existing Remap function
variants have been marked as deprecated in favor of Remap.