on parse, resolve by URN then URI
The strategy here is to leverage the Collection to perform the uri ->
urn resolution. When an extension is encountered in a plan, it checks
for the URN anchor reference, and uses it if it is found. If it is not
found, then it looks for a URI anchor reference and a corresponding
URI -> URN mapping in the collection. If that is also not found, then
we are in an error state.